@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <Version>1.0.1</Version>
     <Description>A small tool to extract xml and pdb files from Nuget Packages.</Description>
     <PackageProjectUrl>https://github.com/guogangj/PdbAndXmlExtractor</PackageProjectUrl>
